requirements.txt
tensorflow_gpu
tifffile
matplotlib
numpy
